apply box details tweaks and file renames
