+++ MTGViewer.Tests/Utils/Dtos/CardDto.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 using System.Text.Json.Serialization;
 using MtgApiManager.Lib.Model;
 
-namespace MTGViewer.Tests.Utils.Dto;
+namespace MTGViewer.Tests.Utils.Dtos;
 
 internal class CardDto
 {
